third_party/blink/web_tests/serial/serialPort_setSignals.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="file:///gen/layout_test_data/mojo/public/js/mojo_bindings.js"></script>
<script src="file:///gen/mojo/public/mojom/base/unguessable_token.mojom.js"></script>
<script src="file:///gen/third_party/blink/public/mojom/serial/serial.mojom.js"></script>
<script src="resources/serial-test-utils.js"></script>
<script>
serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  await promise_rejects(t, 'InvalidStateError', port.setSignals({}));
}, 'setSignals() rejects if the port is not open');

serial_test(async (t, fake) => {
  const { port, fakePort } = await getFakeSerialPort(fake);
  await port.open({ baudrate: 9600 });

  let expectedSignals = { dtr: true, rts: false, brk: false };
  assert_object_equals(fakePort.outputSignals, expectedSignals, "initial");

  await port.setSignals({});
  assert_object_equals(fakePort.outputSignals, expectedSignals, "no-op");

  await port.setSignals({ dtr: false });
  expectedSignals.dtr = false;
  assert_object_equals(fakePort.outputSignals, expectedSignals, "clear DTR");

  await port.setSignals({ rts: true });
  expectedSignals.rts = true;
  assert_object_equals(fakePort.outputSignals, expectedSignals, "set RTS");

  await port.setSignals({ brk: true });
  expectedSignals.brk = true;
  assert_object_equals(fakePort.outputSignals, expectedSignals, "set BRK");

  await port.setSignals({ dtr: true, rts: false, brk: false });
  expectedSignals.dtr = true;
  expectedSignals.rts = false;
  expectedSignals.brk = false;
  assert_object_equals(fakePort.outputSignals, expectedSignals, "invert");
}, 'setSignals() modifies the state of the port');
</script>
